+++ lib/rpmsg_lite/porting/platform/imxrt1160/rpmsg_platform.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2024 NXP
+ * Copyright 2021-2025 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -28,8 +28,12 @@ static LOCK_STATIC_CONTEXT platform_lock_static_ctxt;
 #endif
 
 #if defined(RL_USE_MCMGR_IPC_ISR_HANDLER) && (RL_USE_MCMGR_IPC_ISR_HANDLER == 1)
-static void mcmgr_event_handler(uint16_t vring_idx, void *context)
+static void mcmgr_event_handler(mcmgr_core_t coreNum, uint16_t vring_idx, void *context)
 {
+    /* Unused */
+    (void)context;
+    (void)coreNum;
+
     env_isr((uint32_t)vring_idx);
 }
 #else
@@ -153,7 +157,11 @@ void platform_notify(uint32_t vector_id)
 {
     env_lock_mutex(platform_lock);
 #if defined(RL_USE_MCMGR_IPC_ISR_HANDLER) && (RL_USE_MCMGR_IPC_ISR_HANDLER == 1)
-    (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)RL_GET_Q_ID(vector_id));
+#if defined(FSL_FEATURE_MU_SIDE_A)
+(void)MCMGR_TriggerEventForce(kMCMGR_Core1, kMCMGR_RemoteRPMsgEvent, (uint16_t)RL_GET_Q_ID(vector_id));
+#elif defined(FSL_FEATURE_MU_SIDE_B)
+(void)MCMGR_TriggerEventForce(kMCMGR_Core0, kMCMGR_RemoteRPMsgEvent, (uint16_t)RL_GET_Q_ID(vector_id));
+#endif
 #else
 /* Write directly into the MU Control Register to trigger General Purpose Interrupt Request (GIR).
    No need to wait until the previous interrupt is processed because the same value