+++ lib/rpmsg_lite/porting/platform/imxrt700_hifi1/rpmsg_platform.c
@@ -37,7 +37,7 @@ static LOCK_STATIC_CONTEXT platform_lock_static_ctxt;
 #endif
 
 #if defined(RL_USE_MCMGR_IPC_ISR_HANDLER) && (RL_USE_MCMGR_IPC_ISR_HANDLER == 1)
-static void mcmgr_event_handler(uint16_t vring_idx, void *context, mcmgr_core_t coreNum)
+static void mcmgr_event_handler(mcmgr_core_t coreNum, uint16_t vring_idx, void *context)
 {
     env_isr((uint32_t)vring_idx);
 }
@@ -115,7 +115,7 @@ void platform_notify(uint32_t vector_id)
 {
     env_lock_mutex(platform_lock);
 #if defined(RL_USE_MCMGR_IPC_ISR_HANDLER) && (RL_USE_MCMGR_IPC_ISR_HANDLER == 1)
-    (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core1);
+    (void)MCMGR_TriggerEventForce(kMCMGR_Core1, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #else
     (void)MU_TriggerInterrupts(MU3_MUB, MU_GI_INTR(1UL << (RL_GET_Q_ID(vector_id))));
 #endif