+++ lib/rpmsg_lite/porting/platform/imxrt700_m33/rpmsg_platform.c
@@ -51,7 +51,7 @@ static LOCK_STATIC_CONTEXT platform_lock_static_ctxt;
 #endif
 
 #if defined(RL_USE_MCMGR_IPC_ISR_HANDLER) && (RL_USE_MCMGR_IPC_ISR_HANDLER == 1)
-static void mcmgr_event_handler(uint16_t vring_idx, void *context, mcmgr_core_t coreNum)
+static void mcmgr_event_handler(mcmgr_core_t coreNum, uint16_t vring_idx, void *context)
 {
     env_isr((uint32_t)vring_idx);
 }
@@ -300,26 +300,26 @@ void platform_notify(uint32_t vector_id)
         case RL_PLATFORM_IMXRT700_M33_0_M33_1_COM_ID:
 #if (defined(MIMXRT735S_cm33_core0_SERIES) || defined(MIMXRT758S_cm33_core0_SERIES) || \
      defined(MIMXRT798S_cm33_core0_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core1);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core1, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #elif (defined(MIMXRT735S_cm33_core1_SERIES) || defined(MIMXRT758S_cm33_core1_SERIES) || \
        defined(MIMXRT798S_cm33_core1_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core0);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core0, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #endif
             break;
         case RL_PLATFORM_IMXRT700_M33_0_HIFI4_COM_ID:
 #if (defined(MIMXRT735S_cm33_core0_SERIES) || defined(MIMXRT758S_cm33_core0_SERIES) || \
      defined(MIMXRT798S_cm33_core0_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core2);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core2, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #elif (defined(MIMXRT735S_hifi4_SERIES) || defined(MIMXRT758S_hifi4_SERIES) || defined(MIMXRT798S_hifi4_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core0);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core0, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #endif
             break;
         case RL_PLATFORM_IMXRT700_M33_1_HIFI1_COM_ID:
 #if (defined(MIMXRT735S_cm33_core1_SERIES) || defined(MIMXRT758S_cm33_core1_SERIES) || \
      defined(MIMXRT798S_cm33_core1_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core3);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core3, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #elif (defined(MIMXRT735S_hifi1_SERIES) || defined(MIMXRT758S_hifi1_SERIES) || defined(MIMXRT798S_hifi1_SERIES))
-            (void)MCMGR_TriggerEventForce(kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id, kMCMGR_Core1);
+            (void)MCMGR_TriggerEventForce(kMCMGR_Core1, kMCMGR_RemoteRPMsgEvent, (uint16_t)vector_id);
 #endif
             break;
         default: